casting an accumulator as the same dtype of the output, i.e. if phase errors then complex, otherwise float

+++ src/radartools/amb_func_tools.py
@@ -887,7 +887,7 @@ def alt_code_fraclag(codeset, baud, fraction, h, lags, ph=''):
 
                 decode=decodeset[jj,iii,:]
 
-                wtmp = np.zeros(cwtt[0].shape)
+                wtmp = np.zeros(cwtt[0].shape, dtype=env.dtype)
                 for ii in range(scancount):
                     wtmp+=decode[ii]*cwtt[ii]
